BREAKING CHANGE: Refactor notifyRecipient to notifyPdt
